requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytesseract==0.3.9
requests==2.28.1
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
